added additional person feature, cosmetic only changes
