+++ pyrcamp_app/lib/features/main_screen/main_screen.dart
@@ -1,6 +1,6 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:pyrcamp_app/features/form_screen/form_screen.dart';
+import 'package:pyrcamp_app/features/form_screen/main_person_form.dart';
 import 'package:pyrcamp_app/features/main_screen/widgets/menu_button.dart';
 import 'package:pyrcamp_app/features/main_screen/widgets/features_list.dart';
 import 'package:pyrcamp_app/features/nav_bar_screen/nav_bar_screen.dart';
